chore(data-service): remove NativeClient indirection COMPASS-5029 (#2425)
