+++ packages/compass-import-export/src/utils/collection-stream.js
@@ -42,7 +42,7 @@ class WritableCollectionStream extends Writable {
   }
 
   _collection() {
-    return this.dataService.client._collection(this.ns);
+    return this.dataService._collection(this.ns);
   }
 
   _write(document, _encoding, next) {
