+++ packages/compass-ssh-tunnel-status/src/stores/store.spec.js
@@ -24,12 +24,10 @@ describe('SshTunnelStatusStore [Store]', () => {
   describe('#onConnected', () => {
     context('when the host is less that 25 chars', () => {
       const dataService = {
-        client: {
-          model: {
-            sshTunnel: 'USER_PASSWORD',
-            sshTunnelHostname: '123.45.67.8',
-            sshTunnelOptions: { dstPort: '27019' }
-          }
+        model: {
+          sshTunnel: 'USER_PASSWORD',
+          sshTunnelHostname: '123.45.67.8',
+          sshTunnelOptions: { dstPort: '27019' }
         }
       };
 
@@ -46,12 +44,10 @@ describe('SshTunnelStatusStore [Store]', () => {
 
     context('when teh host is greater than 25 characters', () => {
       const dataService = {
-        client: {
-          model: {
-            sshTunnel: 'USER_PASSWORD',
-            sshTunnelHostname: 'areallylong.amazonlikeaws.com',
-            sshTunnelOptions: { dstPort: '27019' }
-          }
+        model: {
+          sshTunnel: 'USER_PASSWORD',
+          sshTunnelHostname: 'areallylong.amazonlikeaws.com',
+          sshTunnelOptions: { dstPort: '27019' }
         }
       };
