+++ packages/data-service/.mocharc.js
@@ -1,1 +1,7 @@
-module.exports = require('@mongodb-js/mocha-config-compass');
+const preset = require('@mongodb-js/mocha-config-compass');
+
+module.exports = {
+  ...preset,
+  spec: '{src,test}/**/*.spec.*',
+  watchFiles: '{src,test}/**/*',
+};
