+++ packages/data-service/src/data-service.spec.ts
@@ -1,21 +1,99 @@
 import assert from 'assert';
 import { ObjectId } from 'bson';
 import { expect } from 'chai';
+import { EventEmitter } from 'events';
+import { MongoClient, Sort } from 'mongodb';
 import sinon from 'sinon';
 import * as helper from '../test/helper';
 import DataService from './data-service';
+import { Callback } from './types';
+
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const mock = require('mock-require');
+
+function mockedTopologyDescription(
+  topologyType = 'Standalone',
+  serverType = 'Single'
+) {
+  return {
+    type: topologyType,
+    servers: new Map([['127.0.0.1:27017', { type: serverType }]]),
+  };
+}
+
+/*
+ * pretends to be a connection-model providing every function call
+ * required in NativeClient#connect, but returns topology and connection
+ * params of our choice
+ */
+function mockedConnectionModel(
+  topologyDescription?: any,
+  connectionOptions?: any
+) {
+  const _topologyDescription =
+    topologyDescription || mockedTopologyDescription();
+
+  const _connectionOptions = connectionOptions || {
+    url: 'mongodb://127.0.0.1:27018/data-service?readPreference=primary&ssl=false',
+    options: {
+      readPreference: 'primary',
+      useNewUrlParser: true,
+      useUnifiedTopology: true,
+    },
+  };
+
+  const mockedTunnel = {
+    close() {
+      return Promise.resolve();
+    },
+  };
+
+  return function (_model: any, setupListeners: any, cb: any) {
+    const mockedClient = new EventEmitter() as any;
+    mockedClient.db = () => {
+      // pass
+    };
+    mockedClient.close = (_force: any, closeCb: Callback<void>) => {
+      closeCb(null);
+    };
+    setupListeners(mockedClient);
+    mockedClient.emit('topologyDescriptionChanged', {
+      newDescription: _topologyDescription,
+    });
+    cb(null, mockedClient, mockedTunnel, _connectionOptions);
+  };
+}
 
 describe('DataService', function () {
   this.slow(10000);
   this.timeout(20000);
+
   let service: DataService;
+  let mongoClient: MongoClient;
+  let sandbox: sinon.SinonSandbox;
 
   before(function (done) {
+    sandbox = sinon.createSandbox();
     service = new DataService(helper.connection);
-    service.connect(done);
+    service.connect((err) => {
+      if (err) return done(err);
+      const opts = service.getMongoClientConnectionOptions();
+      MongoClient.connect(opts!.url, opts!.options, (err, client) => {
+        if (err) {
+          return done(err);
+        }
+        mongoClient = client!;
+        done();
+      });
+    });
   });
   after(function (done) {
-    service.disconnect(done);
+    service.disconnect(() => {
+      mongoClient.close(done);
+    });
+  });
+  afterEach(function () {
+    sandbox.restore();
   });
 
   // Each test gets its own service so that we can connect/disconnect it freely
@@ -101,7 +179,7 @@ describe('DataService', function () {
 
   describe('#dropCollection', function () {
     before(function (done) {
-      service.client.database.createCollection('bar', {}, function (error) {
+      mongoClient.db().createCollection('bar', {}, function (error) {
         assert.equal(null, error);
         done();
       });
@@ -122,7 +200,7 @@ describe('DataService', function () {
 
   describe('#dropDatabase', function () {
     before(function (done) {
-      service.client.client
+      mongoClient
         .db('mangoDB')
         .createCollection('testing', {}, function (error) {
           assert.equal(null, error);
@@ -144,16 +222,19 @@ describe('DataService', function () {
 
   describe('#dropIndex', function () {
     before(function (done) {
-      service.client.database.collection('test').createIndex(
-        {
-          a: 1,
-        },
-        {},
-        function (error) {
-          assert.equal(null, error);
-          done();
-        }
-      );
+      mongoClient
+        .db()
+        .collection('test')
+        .createIndex(
+          {
+            a: 1,
+          },
+          {},
+          function (error) {
+            assert.equal(null, error);
+            done();
+          }
+        );
     });
 
     it('removes an index from a collection', function (done) {
@@ -208,13 +289,13 @@ describe('DataService', function () {
 
   describe('#aggregate', function () {
     before(function (done) {
-      helper.insertTestDocuments(service.client, function () {
+      helper.insertTestDocuments(mongoClient, function () {
         done();
       });
     });
 
     after(function (done) {
-      helper.deleteTestDocuments(service.client, function () {
+      helper.deleteTestDocuments(mongoClient, function () {
         done();
       });
     });
@@ -257,13 +338,13 @@ describe('DataService', function () {
 
   describe('#find', function () {
     before(function (done) {
-      helper.insertTestDocuments(service.client, function () {
+      helper.insertTestDocuments(mongoClient, function () {
         done();
       });
     });
 
     after(function (done) {
-      helper.deleteTestDocuments(service.client, function () {
+      helper.deleteTestDocuments(mongoClient, function () {
         done();
       });
     });
@@ -282,17 +363,76 @@ describe('DataService', function () {
         }
       );
     });
+
+    context('when a filter is provided', function () {
+      it('returns a cursor for the matching documents', function (done) {
+        service.find(
+          'data-service.test',
+          {
+            a: 1,
+          },
+          {},
+          function (error, docs) {
+            assert.equal(null, error);
+            expect(docs.length).to.equal(1);
+            done();
+          }
+        );
+      });
+    });
+
+    context('when no filter is provided', function () {
+      it('returns a cursor for all documents', function (done) {
+        service.find('data-service.test', {}, {}, function (error, docs) {
+          assert.equal(null, error);
+          expect(docs.length).to.equal(2);
+          done();
+        });
+      });
+    });
+
+    context('when options are provided', function () {
+      it('returns a cursor for the documents', function (done) {
+        service.find(
+          'data-service.test',
+          {},
+          {
+            skip: 1,
+          },
+          function (error, docs) {
+            assert.equal(null, error);
+            expect(docs.length).to.equal(1);
+            done();
+          }
+        );
+      });
+    });
+
+    context('when array sort is provided', function () {
+      it('returns documents with correct sort order', function (done) {
+        const sort: Sort = [
+          ['2', -1],
+          ['1', -1],
+        ];
+        service.find('data-service.test', {}, { sort }, function (error, docs) {
+          assert.strictEqual(null, error);
+          expect(docs[0]).to.have.nested.property('2', 'a');
+          expect(docs[1]).to.have.nested.property('1', 'a');
+          done();
+        });
+      });
+    });
   });
 
   describe('#fetch', function () {
     before(function (done) {
-      helper.insertTestDocuments(service.client, function () {
+      helper.insertTestDocuments(mongoClient, function () {
         done();
       });
     });
 
     after(function (done) {
-      helper.deleteTestDocuments(service.client, function () {
+      helper.deleteTestDocuments(mongoClient, function () {
         done();
       });
     });
@@ -305,11 +445,44 @@ describe('DataService', function () {
         done();
       });
     });
+
+    context('when a filter is provided', function () {
+      it('returns a cursor for the matching documents', function (done) {
+        const cursor = service.fetch('data-service.test', { a: 1 }, {});
+        cursor.toArray(function (error, docs) {
+          assert.equal(null, error);
+          expect(docs!.length).to.equal(1);
+          done();
+        });
+      });
+    });
+
+    context('when no filter is provided', function () {
+      it('returns a cursor for all documents', function (done) {
+        const cursor = service.fetch('data-service.test', {}, {});
+        cursor.toArray(function (error, docs) {
+          assert.equal(null, error);
+          expect(docs!.length).to.equal(2);
+          done();
+        });
+      });
+    });
+
+    context('when options are provided', function () {
+      it('returns a cursor for the documents', function (done) {
+        const cursor = service.fetch('data-service.test', {}, { skip: 1 });
+        cursor.toArray(function (error, docs) {
+          assert.equal(null, error);
+          expect(docs!.length).to.equal(1);
+          done();
+        });
+      });
+    });
   });
 
   describe('#findOneAndReplace', function () {
     after(function (done) {
-      helper.deleteTestDocuments(service.client, function () {
+      helper.deleteTestDocuments(mongoClient, function () {
         done();
       });
     });
@@ -352,7 +525,7 @@ describe('DataService', function () {
 
   describe('#findOneAndUpdate', function () {
     after(function (done) {
-      helper.deleteTestDocuments(service.client, function () {
+      helper.deleteTestDocuments(mongoClient, function () {
         done();
       });
     });
@@ -469,6 +642,32 @@ describe('DataService', function () {
         );
       });
     });
+
+    context('when max timeout is provided', function () {
+      context('when the count times out', function () {
+        before(function (done) {
+          service.insertOne('data-service.test', { a: 500 }, {}, done);
+        });
+
+        after(function (done) {
+          service.deleteMany('data-service.test', {}, {}, done);
+        });
+
+        it('does not throw the error', function (done) {
+          service.count(
+            'data-service.test',
+            {
+              $where: 'function() { sleep(5500); return true; }',
+            },
+            { maxTimeMS: 5000 },
+            function (error) {
+              expect(error).to.not.equal(null);
+              done();
+            }
+          );
+        });
+      });
+    });
   });
 
   describe('#database', function () {
@@ -487,7 +686,7 @@ describe('DataService', function () {
       service.listDatabases(function (err, databases) {
         assert.equal(null, err);
         const databaseNames = databases.map((db: any) => db.name);
-        if (service.client.isMongos) {
+        if (service.isMongos()) {
           expect(databaseNames).to.not.contain('local');
         } else {
           expect(databaseNames).to.contain('local');
@@ -502,7 +701,7 @@ describe('DataService', function () {
 
   describe('#createCollection', function () {
     after(function (done) {
-      service.client.database.dropCollection('foo', {}, function (error) {
+      mongoClient.db().dropCollection('foo', {}, function (error) {
         assert.equal(null, error);
         done();
       });
@@ -538,7 +737,7 @@ describe('DataService', function () {
 
   describe('#createIndex', function () {
     after(function (done) {
-      const testCollection = service.client.database.collection('test');
+      const testCollection = mongoClient.db().collection('test');
       testCollection.dropIndex('a_1', {}, function () {
         testCollection.dropIndex('b_1', {}, function () {
           testCollection.dropIndex('a_-1_b_1', {}, function () {
@@ -627,7 +826,7 @@ describe('DataService', function () {
 
   describe('#insertOne', function () {
     after(function (done) {
-      helper.deleteTestDocuments(service.client, function () {
+      helper.deleteTestDocuments(mongoClient, function () {
         done();
       });
     });
@@ -660,7 +859,7 @@ describe('DataService', function () {
 
   describe('#insertMany', function () {
     after(function (done) {
-      helper.deleteTestDocuments(service.client, function () {
+      helper.deleteTestDocuments(mongoClient, function () {
         done();
       });
     });
@@ -697,45 +896,111 @@ describe('DataService', function () {
   });
 
   describe('#sample', function () {
-    it('passes arguments to native client', function () {
-      const sampleSpy = sinon.spy(() => 'expectedResult');
-      const dataService = new DataService(helper.connection);
-      dataService.client = { sample: sampleSpy } as any;
-      const res = dataService.sample('a.b', { size: 1 }, { batchSize: 10 });
-      expect(res).to.equal('expectedResult');
-      expect(sampleSpy).to.have.been.calledOnceWith(
-        'a.b',
-        { size: 1 },
-        { batchSize: 10 }
+    before(function (done) {
+      helper.insertTestDocuments(mongoClient, function () {
+        done();
+      });
+    });
+
+    after(function (done) {
+      helper.deleteTestDocuments(mongoClient, function () {
+        done();
+      });
+    });
+
+    it('returns a cursor of sampled documents', async function () {
+      const docs = await service.sample('data-service.test').toArray();
+      expect(docs.length).to.equal(2);
+    });
+
+    it('allows to pass a query', async function () {
+      const docs = await service
+        .sample('data-service.test', {
+          query: { a: 1 },
+        })
+        .toArray();
+      expect(docs.length).to.equal(1);
+      expect(docs[0]).to.haveOwnProperty('_id');
+      expect(docs[0].a).to.equal(1);
+    });
+
+    it('allows to pass a projection', async function () {
+      const docs = await service
+        .sample('data-service.test', {
+          fields: {
+            a: 1,
+            _id: 0,
+          },
+        })
+        .toArray();
+
+      expect(docs).to.deep.include.members([{ a: 1 }, { a: 2 }]);
+    });
+
+    it('allows to set a sample size', async function () {
+      const docs = await service
+        .sample('data-service.test', {
+          size: 1,
+        })
+        .toArray();
+
+      expect(docs.length).to.equal(1);
+    });
+
+    it('always sets default sample size and allowDiskUse: true', function () {
+      sandbox.spy(service, 'aggregate');
+      service.sample('db.coll');
+
+      // eslint-disable-next-line @typescript-eslint/unbound-method
+      expect(service.aggregate).to.have.been.calledWith(
+        'db.coll',
+        [{ $sample: { size: 1000 } }],
+        { allowDiskUse: true }
       );
     });
-  });
 
-  describe('#startSession', function () {
-    it('passes arguments to native client', function () {
-      const startSessionSpy = sinon.spy(() => 'expectedResult');
-      const dataService = new DataService(helper.connection);
-      dataService.client = { startSession: startSessionSpy } as any;
-      const res = dataService.startSession();
-      expect(res).to.equal('expectedResult');
-      expect(startSessionSpy).to.have.been.calledOnceWith();
+    it('allows to pass down aggregation options to the driver', function () {
+      sandbox.spy(service, 'aggregate');
+      service.sample(
+        'db.coll',
+        {},
+        {
+          maxTimeMS: 123,
+          session: undefined,
+          raw: true,
+        }
+      );
+
+      // eslint-disable-next-line @typescript-eslint/unbound-method
+      expect(service.aggregate).to.have.been.calledWith(
+        'db.coll',
+        [{ $sample: { size: 1000 } }],
+        { allowDiskUse: true, maxTimeMS: 123, session: undefined, raw: true }
+      );
     });
-  });
 
-  describe('#killSession', function () {
-    it('passes arguments to native client', function () {
-      const killSessionSpy = sinon.spy(() => 'expectedResult');
-      const dataService = new DataService(helper.connection);
-      dataService.client = { killSession: killSessionSpy } as any;
-      const res = dataService.killSession({ id: 123 } as any);
-      expect(res).to.equal('expectedResult');
-      expect(killSessionSpy).to.have.been.calledOnceWith({ id: 123 });
+    it('allows to override allowDiskUse', function () {
+      sandbox.spy(service, 'aggregate');
+      service.sample(
+        'db.coll',
+        {},
+        {
+          allowDiskUse: false,
+        }
+      );
+
+      // eslint-disable-next-line @typescript-eslint/unbound-method
+      expect(service.aggregate).to.have.been.calledWith(
+        'db.coll',
+        [{ $sample: { size: 1000 } }],
+        { allowDiskUse: false }
+      );
     });
   });
 
   describe('#updateOne', function () {
     after(function (done) {
-      helper.deleteTestDocuments(service.client, function () {
+      helper.deleteTestDocuments(mongoClient, function () {
         done();
       });
     });
@@ -806,7 +1071,7 @@ describe('DataService', function () {
 
   describe('#updateMany', function () {
     after(function (done) {
-      helper.deleteTestDocuments(service.client, function () {
+      helper.deleteTestDocuments(mongoClient, function () {
         done();
       });
     });
@@ -859,13 +1124,13 @@ describe('DataService', function () {
 
   describe('#views', function () {
     before(function (done) {
-      helper.insertTestDocuments(service.client, function () {
+      helper.insertTestDocuments(mongoClient, function () {
         done();
       });
     });
 
     after(function (done) {
-      helper.deleteTestDocuments(service.client, function () {
+      helper.deleteTestDocuments(mongoClient, function () {
         done();
       });
     });
@@ -928,6 +1193,309 @@ describe('DataService', function () {
 
         assert.equal(_count, 0);
         done();
+      });
+    });
+  });
+
+  describe('#connect', function () {
+    context('when mocking connection-model', function () {
+      after(function () {
+        mock.stop('mongodb-connection-model');
+      });
+
+      it('does not allow to connect twice without disonnecting first', function (done) {
+        mock('./legacy-connect', mockedConnectionModel());
+
+        const MockedDataService = mock.reRequire('./data-service');
+        const mockedService: DataService = new MockedDataService(
+          helper.connection
+        );
+
+        mockedService.connect(() => {
+          // pass
+        });
+        mockedService.connect((err: any) => {
+          expect(err).to.be.instanceOf(Error);
+          expect(err)
+            .to.have.property('message')
+            .match(
+              /Connect method has been called more than once without disconnecting/
+            );
+
+          done();
+        });
+      });
+
+      it('sets .connectionOptions after successful connection', function (done) {
+        mock('./legacy-connect', mockedConnectionModel());
+
+        const MockedDataService = mock.reRequire('./data-service');
+        const mockedService: DataService = new MockedDataService(
+          helper.connection
+        );
+
+        expect(mockedService.getMongoClientConnectionOptions()).to.be.undefined;
+
+        mockedService.connect(function () {
+          expect(mockedService.getMongoClientConnectionOptions()).to.deep.equal(
+            {
+              url: 'mongodb://127.0.0.1:27018/data-service?readPreference=primary&ssl=false',
+              options: {
+                readPreference: 'primary',
+                useNewUrlParser: true,
+                useUnifiedTopology: true,
+              },
+            }
+          );
+          done();
+        });
+      });
+
+      it('sets .isMongos to true when topology is sharded', function (done) {
+        mock(
+          './legacy-connect',
+          mockedConnectionModel(mockedTopologyDescription('Sharded'))
+        );
+
+        const MockedDataService = mock.reRequire('./data-service');
+        const mockedService: DataService = new MockedDataService(
+          helper.connection
+        );
+
+        mockedService.connect(function () {
+          expect(mockedService.isMongos()).to.be.true;
+          done();
+        });
+      });
+
+      it('sets .isMongos to false when topology is not sharded', function (done) {
+        mock('./legacy-connect', mockedConnectionModel());
+
+        const MockedDataService = mock.reRequire('./data-service');
+        const mockedService: DataService = new MockedDataService(
+          helper.connection
+        );
+
+        mockedService.connect(function () {
+          expect(mockedService.isMongos()).to.be.false;
+          done();
+        });
+      });
+    });
+  });
+
+  describe('#disconnect', function () {
+    context('when mocking connection-model', function () {
+      after(function () {
+        mock.stop('mongodb-connection-model');
+      });
+
+      it('should close tunnel before calling disconnect callback', function (done) {
+        mock('./legacy-connect', mockedConnectionModel());
+
+        const MockedDataService = mock.reRequire('./data-service');
+        const mockedService: DataService = new MockedDataService(
+          helper.connection
+        );
+
+        mockedService.connect(() => {
+          const closeSpy = sandbox.spy((mockedService as any)._tunnel, 'close');
+
+          const disconnectCallbackSpy = sandbox.spy(() => {
+            try {
+              expect(closeSpy).to.have.been.calledOnce;
+              expect(closeSpy).to.have.been.calledBefore(disconnectCallbackSpy);
+              done();
+            } catch (err) {
+              done(err);
+            }
+          });
+
+          mockedService.disconnect(disconnectCallbackSpy);
+        });
+      });
+    });
+  });
+
+  // describe('#collectionDetail', function () {
+  //   it('returns the collection details', function (done) {
+  //     service._collectionDetail('data-service.test', function (err, coll) {
+  //       assert.equal(null, err);
+  //       expect(coll.ns).to.equal('data-service.test');
+  //       expect(coll.index_count).to.equal(1);
+  //       done();
+  //     });
+  //   });
+  // });
+
+  // describe('#collectionNames', function () {
+  //   it('returns the collection names', function (done) {
+  //     service._collectionNames('data-service', function (err, names) {
+  //       assert.equal(null, err);
+  //       expect(names[0]).to.not.equal(undefined);
+  //       done();
+  //     });
+  //   });
+  // });
+
+  describe('#collections', function () {
+    context('when no readonly views exist', function () {
+      it('returns the collections', function (done) {
+        service.collections('data-service', function (err, collections) {
+          assert.equal(null, err);
+          expect(collections[0].name).to.not.equal(undefined);
+          done();
+        });
+      });
+    });
+
+    context('when readonly views exist', function () {
+      after(function (done) {
+        service.dropCollection('data-service.readonlyfoo', function () {
+          service.dropCollection('data-service.system.views', function () {
+            done();
+          });
+        });
+      });
+
+      it('returns empty stats for the readonly views', function (done) {
+        const pipeline = [{ $match: { name: 'test' } }];
+        const options = { viewOn: 'test', pipeline: pipeline };
+        service.createCollection(
+          'data-service.readonlyfoo',
+          options,
+          function (error) {
+            if (error) {
+              assert.notEqual(null, error.message);
+              done();
+            } else {
+              service.collections('data-service', function (err, collections) {
+                assert.equal(null, err);
+                expect(collections[0].name).to.not.equal(undefined);
+                done();
+              });
+            }
+          }
+        );
+      });
+    });
+  });
+
+  describe('#currentOp', function () {
+    it('returns an object with the currentOp', function (done) {
+      service.currentOp(true, function (err, result) {
+        assert.equal(null, err);
+        expect(result.inprog).to.not.equal(undefined); // TODO: are these tests enough?
+        done();
+      });
+    });
+  });
+
+  describe('#serverstats', function () {
+    it('returns an object with the serverstats', function (done) {
+      service.serverstats(function (err, result) {
+        assert.equal(null, err);
+        expect(result.ok).to.equal(1);
+        done();
+      });
+    });
+  });
+
+  describe('#top', function () {
+    it('returns an object with the results from top', function (done) {
+      service.top(function (err, result) {
+        if (service.isMongos()) {
+          assert(err);
+          expect(err.message).to.contain('top');
+          done();
+          return;
+        }
+        assert.equal(null, err);
+        expect(result.ok).to.equal(1);
+        done();
+      });
+    });
+  });
+
+  // describe('#databaseDetail', function () {
+  //   it('returns the database details', function (done) {
+  //     service.databaseDetail('data-service', function (err, database) {
+  //       assert.equal(null, err);
+  //       expect(database._id).to.equal('data-service');
+  //       expect(database.stats.document_count).to.not.equal(undefined);
+  //       done();
+  //     });
+  //   });
+  // });
+
+  // describe('#databaseStats', function () {
+  //   context('when the user is authorized', function () {
+  //     it('returns an object with the db stats', function (done) {
+  //       service.databaseStats('native-service', function (err, stats) {
+  //         assert.equal(null, err);
+  //         expect(stats.document_count).to.equal(0);
+  //         done();
+  //       });
+  //     });
+  //   });
+
+  //   context('when the user is not authorized', function () {
+  //     it('passes an error to the callback');
+  //   });
+  // });
+
+  describe('#explain', function () {
+    context('when a filter is provided', function () {
+      it('returns an explain object for the provided filter', function (done) {
+        service.explain(
+          'data-service.test',
+          {
+            a: 1,
+          },
+          {},
+          function (error, explanation) {
+            assert.equal(null, error);
+            expect(explanation).to.be.an('object');
+            done();
+          }
+        );
+      });
+    });
+  });
+
+  describe('#startSession', function () {
+    it('returns a new client session', function () {
+      const session = service.startSession();
+      expect(session.constructor.name).to.equal('ClientSession');
+
+      // used by killSession, must be a bson UUID in order to work
+      expect(session.id!.id._bsontype).to.equal('Binary');
+      expect(session.id!.id.sub_type).to.equal(4);
+    });
+  });
+
+  describe('#killSession', function () {
+    it('does not throw if kill a non existing session', async function () {
+      const session = service.startSession();
+      await service.killSession(session);
+    });
+
+    it('kills a command with a session', async function () {
+      const commandSpy = sinon.spy();
+      sandbox.replace(
+        (service as any)._client,
+        'db',
+        () =>
+          ({
+            command: commandSpy,
+          } as any)
+      );
+
+      const session = service.startSession();
+      await service.killSession(session);
+
+      expect(commandSpy.args[0][0]).to.deep.equal({
+        killSessions: [session.id],
       });
     });
   });
